requirements.txt
flask
gunicorn
huggingface_hub
python-dotenv
requests

